added xml and readme images
