+++ source/Kangaroo/Platforms/WindowsArp.cs
@@ -1,9 +1,8 @@
-﻿namespace Kangaroo.Platforms
-{
-    internal class WindowsArp
-    {
-        [System.Runtime.InteropServices.DllImport("iphlpapi.dll", EntryPoint = "SendARP")]
-        internal static extern int SendARP(int destIp, int srcIp, byte[] macAddr, ref uint physicalAddrLen);
+﻿namespace Kangaroo.Platforms;
 
-    }
+internal class WindowsArp
+{
+    [System.Runtime.InteropServices.DllImport("iphlpapi.dll", EntryPoint = "SendARP")]
+    internal static extern int SendARP(int destIp, int srcIp, byte[] macAddr, ref uint physicalAddrLen);
 }
+
