+++ source/Kangaroo/Queries/IQueryPingResults.cs
@@ -1,6 +1,4 @@
-﻿
-
-using System.Net;
+﻿using System.Net;
 using System.Net.NetworkInformation;
 
 namespace Kangaroo.Queries;
